refactor: Hero 컴포넌트에서 hook분리를 통해 추상화
