+++ client/src/shared/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useDelayedVisible } from './useDelayedVisible';
