+++ client/src/widgets/Hero/Hero.styles.ts
